pyproject.toml
[project]
name = "mcp-server-github"
version = "0.1.0"
description = "A FastAPI-based MCP server for GitHub automation"
authors = [
    { name="Your Name", email="your@email.com" }
]
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "PyGithub",
    "python-dotenv"
]
requires-python = ">=3.8"

[tool.uv]
# Configuration for uv if needed 